image download ok

+++ src/main/java/com/raina/traduler/themeList/dto/ThemeListResponse.java
@@ -15,7 +15,7 @@ import java.util.stream.Collectors;
 public class ThemeListResponse {
 
     /*ThemeListEntity*/
-    private int placeId;
+    private Long placeId;
     private String placeName;
     private String contentIntro;
     private String contentMain;
